Vistas del admin
Completo las funcionalidades , falta notificar que se modifico la sesion

+++ ProyectoProgra1/src/server/view/AdminView.java
@@ -20,8 +20,8 @@ public class AdminView extends javax.swing.JFrame {
     /**
      * Probando pull Creates new form AdminView
      */
-    private SessionListModel sessionModel = new SessionListModel();
-    private UserListModel userModel = new UserListModel();
+    private SessionListModel sessionModel;
+    private UserListModel userModel;
     private Data data;
 
     public AdminView(javax.swing.JFrame parent, boolean modal, Data data) {
@@ -31,6 +31,7 @@ public class AdminView extends javax.swing.JFrame {
         jListArea.setVisible(false);
         btnDeleteSession.setVisible(false);
         btnModify.setVisible(false);
+        btnSessionDetails.setVisible(false);
         this.data = data;
 
     }
@@ -52,6 +53,7 @@ public class AdminView extends javax.swing.JFrame {
         jListArea = new javax.swing.JList();
         btnDeleteSession = new javax.swing.JButton();
         btnModify = new javax.swing.JButton();
+        btnSessionDetails = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("Administrador");
@@ -119,6 +121,13 @@ public class AdminView extends javax.swing.JFrame {
             }
         });
 
+        btnSessionDetails.setText("Ver Sesion");
+        btnSessionDetails.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnSessionDetailsActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -134,16 +143,18 @@ public class AdminView extends javax.swing.JFrame {
                     .addComponent(btnShowSessions, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(btnCreateSession, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(btnManageSessions, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 83, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 87, Short.MAX_VALUE)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addComponent(btnDeleteSession, javax.swing.GroupLayout.PREFERRED_SIZE, 99, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(btnDeleteSession, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(15, 15, 15)
+                        .addComponent(btnModify, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(btnModify, javax.swing.GroupLayout.PREFERRED_SIZE, 99, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(124, 124, 124))
+                        .addComponent(btnSessionDetails, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(103, 103, 103))
                     .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
                         .addComponent(jListArea, javax.swing.GroupLayout.PREFERRED_SIZE, 432, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(80, 80, 80))))
+                        .addGap(76, 76, 76))))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -156,18 +167,20 @@ public class AdminView extends javax.swing.JFrame {
                         .addComponent(btnManageSessions, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 30, Short.MAX_VALUE)
                         .addComponent(btnCreateSession, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 31, Short.MAX_VALUE)
-                        .addComponent(btnShowSessions, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 30, Short.MAX_VALUE)
+                        .addComponent(btnShowSessions, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 30, Short.MAX_VALUE)
+                        .addComponent(btnShowUsersList, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addContainerGap(61, Short.MAX_VALUE))
                     .addGroup(layout.createSequentialGroup()
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGap(31, 31, 31)
                         .addComponent(jListArea, javax.swing.GroupLayout.PREFERRED_SIZE, 285, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(btnDeleteSession, javax.swing.GroupLayout.DEFAULT_SIZE, 59, Short.MAX_VALUE)
-                            .addComponent(btnModify, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 39, Short.MAX_VALUE)
-                .addComponent(btnShowUsersList, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(45, Short.MAX_VALUE))
+                        .addGap(30, 30, 30)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(btnDeleteSession, javax.swing.GroupLayout.PREFERRED_SIZE, 60, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(btnModify, javax.swing.GroupLayout.PREFERRED_SIZE, 60, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(btnSessionDetails, javax.swing.GroupLayout.PREFERRED_SIZE, 60, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
         );
 
         pack();
@@ -180,6 +193,7 @@ public class AdminView extends javax.swing.JFrame {
     }//GEN-LAST:event_btnCreateSessionActionPerformed
 
     private void btnManageSessionsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnManageSessionsActionPerformed
+        sessionModel = new SessionListModel();
         jListArea.setModel(sessionModel);
         jListArea.setVisible(true);
         for (HashMap.Entry<String, Session> session : data.getSessions().entrySet()) {
@@ -188,31 +202,45 @@ public class AdminView extends javax.swing.JFrame {
         }
         btnDeleteSession.setVisible(true);
         btnModify.setVisible(true);
+        btnSessionDetails.setVisible(false);
 
     }//GEN-LAST:event_btnManageSessionsActionPerformed
 
     private void btnShowSessionsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnShowSessionsActionPerformed
+        sessionModel = new SessionListModel();
         jListArea.setModel(sessionModel);
         jListArea.setVisible(true);
         for (HashMap.Entry<String, Session> session : data.getSessions().entrySet()) {
 
             sessionModel.addSessio(session.getValue());
         }
+        btnDeleteSession.setVisible(false);
+        btnModify.setVisible(false);
+        btnSessionDetails.setVisible(true);
 
     }//GEN-LAST:event_btnShowSessionsActionPerformed
 
     private void btnShowUsersListActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnShowUsersListActionPerformed
+        userModel = new UserListModel();
         jListArea.setModel(userModel);
         jListArea.setVisible(true);
         for (HashMap.Entry<String, User> user : data.getUsers().entrySet()) {
             userModel.addUser(user.getValue());
         }
+        btnDeleteSession.setVisible(false);
+        btnModify.setVisible(false);
+        btnSessionDetails.setVisible(false);
     }//GEN-LAST:event_btnShowUsersListActionPerformed
 
     private void btnModifyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnModifyActionPerformed
-        int index = jListArea.getSelectedIndex();
+        int index = -2;
+        index = jListArea.getSelectedIndex();
         if (index == -1) {
-            JOptionPane.showMessageDialog(this, "No hay sesiones en la lista", "Error", NORMAL);
+            JOptionPane.showMessageDialog(this, "No ha seleccionado una sesion o no hay sesiones en la lista", "Error", NORMAL);
+            return;
+        }
+        if (index == -2) {
+            JOptionPane.showMessageDialog(this, "No se ha seleccionado ninguna sesion", "Error", NORMAL);
             return;
         }
         Session session = sessionModel.getSession(index);
@@ -224,46 +252,38 @@ public class AdminView extends javax.swing.JFrame {
     }//GEN-LAST:event_btnModifyActionPerformed
 
     private void btnDeleteSessionActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteSessionActionPerformed
-        int index = jListArea.getSelectedIndex();
+        int index = -2;
+        index = jListArea.getSelectedIndex();
+        if (index == -1) {
+            JOptionPane.showMessageDialog(this, "No ha seleccionado una sesion o no hay sesiones en la lista", "Error", NORMAL);
+            return;
+        }
+        if (index == -2) {
+            JOptionPane.showMessageDialog(this, "No se ha seleccionado ninguna sesion", "Error", NORMAL);
+            return;
+        }
         data.deleteSession(sessionModel.getSession(index).getSesionId());
         sessionModel.deleteSession(index);
 
     }//GEN-LAST:event_btnDeleteSessionActionPerformed
 
-    /**
-     * @param args the command line arguments
-     */
-//    public static void main(String args[]) {
-//        /* Set the Nimbus look and feel */
-//        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-//        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-//         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-//         */
-//        try {
-//            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-//                if ("Nimbus".equals(info.getName())) {
-//                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-//                    break;
-//                }
-//            }
-//        } catch (ClassNotFoundException ex) {
-//            java.util.logging.Logger.getLogger(AdminView.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (InstantiationException ex) {
-//            java.util.logging.Logger.getLogger(AdminView.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (IllegalAccessException ex) {
-//            java.util.logging.Logger.getLogger(AdminView.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-//            java.util.logging.Logger.getLogger(AdminView.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        }
-//        //</editor-fold>
-//
-//        /* Create and display the form */
-//        java.awt.EventQueue.invokeLater(new Runnable() {
-//            public void run() {
-//                new AdminView().setVisible(true);
-//            }
-//        });
-//    }
+    private void btnSessionDetailsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSessionDetailsActionPerformed
+       int index = -2;
+        index = jListArea.getSelectedIndex();
+        if (index == -1) {
+            JOptionPane.showMessageDialog(this, "No ha seleccionado una sesion o no hay sesiones en la lista", "Error", NORMAL);
+            return;
+        }
+        if (index == -2) {
+            JOptionPane.showMessageDialog(this, "No se ha seleccionado ninguna sesion", "Error", NORMAL);
+            return;
+        }
+        Session session = sessionModel.getSession(index);
+        
+        SessionDetail sessionDetail = new SessionDetail(this, true, data, session);
+        sessionDetail.setVisible(true);
+    }//GEN-LAST:event_btnSessionDetailsActionPerformed
+
     private void views() {
         try {
             for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
@@ -295,6 +315,7 @@ public class AdminView extends javax.swing.JFrame {
     private javax.swing.JButton btnDeleteSession;
     private javax.swing.JButton btnManageSessions;
     private javax.swing.JButton btnModify;
+    private javax.swing.JButton btnSessionDetails;
     private javax.swing.JButton btnShowSessions;
     private javax.swing.JButton btnShowUsersList;
     private javax.swing.JLabel jLabel1;
